refactor(docs): restructure Getting Started page and add Quick Start component
