+++ app/page.tsx
@@ -14,7 +14,7 @@ export default function Home() {
         A set of 8-bit styled components and a code distribution platform. Works
         with your favorite frameworks. Open Source. Open Code.
       </p>
-      <Link href="/docs/getting-started" className="w-fit">
+      <Link href="/docs" className="w-fit">
         <Button size="sm" className="w-fit">
           Get Started
         </Button>
